+++ Amplitude/AMPDatabaseHelper.h
@@ -8,7 +8,10 @@
 
 @interface AMPDatabaseHelper : NSObject
 
+@property (nonatomic, strong, readonly) NSString *databasePath;
+
 + (AMPDatabaseHelper*)getDatabaseHelper;
++ (AMPDatabaseHelper*)getDatabaseHelper:(NSString*) instanceName;
 - (BOOL)createTables;
 - (BOOL)dropTables;
 - (BOOL)upgrade:(int) oldVersion newVersion:(int) newVersion;