+++ Amplitude/Amplitude.h
@@ -40,6 +40,8 @@
 @property (nonatomic, strong, readonly) NSString *apiKey;
 @property (nonatomic, strong, readonly) NSString *userId;
 @property (nonatomic, strong, readonly) NSString *deviceId;
+@property (nonatomic ,strong, readonly) NSString *instanceName;
+@property (nonatomic ,strong, readonly) NSString *propertyListPath;
 @property (nonatomic, assign) BOOL optOut;
 
 /*!
@@ -81,6 +83,8 @@
 #pragma mark - Methods
 
 + (Amplitude *)instance;
+
++ (Amplitude *)instanceWithName:(NSString*) instanceName;
 
 /*!
  @method
@@ -201,6 +205,15 @@
  @method
 
  @abstract
+ Clears all properties that are tracked on the user level.
+ */
+
+- (void)clearUserProperties;
+
+/*!
+ @method
+
+ @abstract
  Sets the userId.
 
  @param userId                  If your app has its own login system that you want to track users with, you can set the userId.
@@ -232,7 +245,7 @@
  @param enabled                  Whether tracking opt out should be enabled or disabled.
 
  @discussion
- If the user wants to opt out of all tracking, use this method to enable opt out for them. Once opt out is enabled, no events will be saved locally or sent to the server. Calling this method again with enabled set to false will turn tracking back on for the user.
+ If the user wants to opt out of all tracking, use this method to enable opt out for them. Once opt out is enabled, no events will be saved locally or sent to the server. Calling this method again with enabled set to NO will turn tracking back on for the user.
  */
 - (void)setOptOut:(BOOL)enabled;
 
@@ -245,7 +258,7 @@
  @param offline                  Whether logged events should be sent to Amplitude servers.
 
  @discussion
- If you want to stop logged events from being sent to Amplitude severs, use this method to set the client to offline. Once offline is enabled, logged events will not be sent to the server until offline is disabled. Calling this method again with offline set to false will allow events to be sent to server
+ If you want to stop logged events from being sent to Amplitude severs, use this method to set the client to offline. Once offline is enabled, logged events will not be sent to the server until offline is disabled. Calling this method again with offline set to NO will allow events to be sent to server
      and the client will attempt to send events that have been queued while offline.
  */
 - (void)setOffline:(BOOL)offline;
@@ -272,6 +285,17 @@
  If you want location tracking disabled on startup of the app, call disableLocationListening before you call initializeApiKey.
  */
 - (void)disableLocationListening;
+
+/*!
+ @method
+
+ @abstract
+ Forces the SDK to update with the user's last known location if possible.
+
+ @discussion
+ If you want to manually force the SDK to update with the user's last known location, call updateLocation.
+ */
+- (void)updateLocation;
 
 /*!
  @method