+++ CHANGELOG.md
@@ -1,6 +1,25 @@
 ## Unreleased
 
+* Add support for logging events to multiple Amplitude apps. See [Readme](https://github.com/amplitude/Amplitude-iOS#tracking-events-to-multiple-amplitude-apps) for details.
+
+## 3.5.0 (January 15, 2016)
+
+* Add ability to clear all user properties.
+
+## 3.4.1 (December 31, 2015)
+
+* Guarding AMPDatabaseHelper logging with a debug flag.
+
+## 3.4.0 (December 29, 2015)
+
+* Remove dependency on FMDB, use built-in SQLite3 library.
+* Updated DeviceInfo platform strings, added iPhone 6s, iPhone 6s Plus, iPod Touch 6G.
+* Fixed bug to make sure events can be serialized before saving.
+
+## 3.3.0 (December 15, 2015)
+
 * Add support for append user property operation.
+* Add ability to force the SDK to update with the user's latest location.
 
 ## 3.2.1 (November 11, 2015)